feat: surface environment badges in zone header
